+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # WIP
+* i : Minimum version Firefox 142
 * Fix : Remove some IDE related warnings
 * \- : Remove `*.prod.` system in the release script and remove `klaw-sync` dependency